requirements.txt
pillow==11.0.0
python-docx==1.1.2
PyPDF2==3.0.1
pdf2image==1.17.0
